Move the decoder inside SierraRecordNumber
