+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/platform/transformer/source/SierraBibData.scala
@@ -1,7 +1,5 @@
 package uk.ac.wellcome.platform.transformer.source
 
-import cats.syntax.either._
-import io.circe.Decoder
 import uk.ac.wellcome.models.transformable.sierra.SierraRecordNumber
 import uk.ac.wellcome.platform.transformer.source.sierra.{
   Country => SierraCountry,
@@ -20,16 +18,3 @@ case class SierraBibData(
   varFields: List[VarField] = List()
 )
 
-/** A Circe decoder that allows unpacking the ID as an instance of
-  * [[SierraRecordNumber]].  This is based on the example Instant decoder
-  * described in the Circe docs.
-  * See: https://circe.github.io/circe/codecs/custom-codecs.html
-  */
-case object SierraBibData {
-  implicit val sierraRecordNumberDecoder: Decoder[SierraRecordNumber] =
-    Decoder.decodeString.emap { str =>
-      Either
-        .catchNonFatal(SierraRecordNumber(str))
-        .leftMap(t => "SierraRecordNumber")
-    }
-}
