+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/platform/transformer/transformers/sierra/SierraItems.scala
@@ -3,13 +3,13 @@ package uk.ac.wellcome.platform.transformer.transformers.sierra
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.models.transformable.SierraTransformable
 import uk.ac.wellcome.models.transformable.sierra.SierraRecordTypes
+import uk.ac.wellcome.models.transformable.sierra.SierraRecordNumber._
 import uk.ac.wellcome.models.work.internal._
 import uk.ac.wellcome.platform.transformer.source.{
   SierraBibData,
   SierraItemData,
   SierraMaterialType
 }
-import uk.ac.wellcome.platform.transformer.source.SierraBibData._
 import uk.ac.wellcome.utils.JsonUtil._
 
 import scala.util.{Failure, Success}
