+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/platform/transformer/source/SierraBibDataTest.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.platform.transformer.source
 
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.transformable.sierra.test.utils.SierraUtil
-import uk.ac.wellcome.platform.transformer.source.SierraBibData._
+import uk.ac.wellcome.models.transformable.sierra.SierraRecordNumber._
 import uk.ac.wellcome.utils.JsonUtil._
 
 class SierraBibDataTest extends FunSpec with Matchers with SierraUtil {
